requirements.txt
streamlit==1.35.0
fastapi==0.112.2
pydantic==1.10.9
uvicorn==0.30.6
mysql-connector-python==8.0.33
pandas==2.0.2
requests==2.31.0
pytest==8.3.2
python-dotenv==1.0.0
